fix bug, system returning fees exceeded when max fee equals fee amount

+++ fio.contracts/contracts/fio.system/include/fio.system/fio.system.hpp
@@ -598,7 +598,7 @@ public:
 
     [[eosio::action]]
     void regproducer(const string fio_address, const std::string &url, uint16_t location, const name actor,
-                     uint16_t max_fee);
+                     uint64_t max_fee);
 
     [[eosio::action]]
     void unregprod(const name producer);
